Fixing Kobe's Shit

+++ Assets/Scripts/EnemyDMG.cs
@@ -6,14 +6,14 @@ public class EnemyDMG : EnemyMind
 {
     // Start is called before the first frame update
     
-    public int enemy_health;
+    public float enemy_health;
     public HealthBar enemy_health_bar;
 
     void Awake() {
         enemy_health_bar.SetMaxHealth(enemy_health);
     }
 
-    public void TakeDMG(int dmg) {
+    public void TakeDMG(float dmg) {
        
     	enemy_health = enemy_health - dmg;
         enemy_health_bar.SetHealth(enemy_health);
